fixes in redo
