+++ app/src/org/commcare/gis/BaseKujakuActivity.kt
@@ -5,12 +5,11 @@ import com.mapbox.mapboxsdk.Mapbox
 import com.mapbox.mapboxsdk.maps.MapboxMap
 import io.ona.kujaku.KujakuLibrary
 import kotlinx.android.synthetic.main.activity_entity_kujaku_map.*
-import kotlinx.coroutines.*
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.cancel
 import org.commcare.activities.CommCareActivity
 import org.commcare.dalvik.BuildConfig
-import org.commcare.dalvik.R
 import org.commcare.interfaces.CommCareActivityUIController
-import org.commcare.interfaces.WithUIController
 
 abstract class BaseKujakuActivity : CommCareActivity<BaseKujakuActivity>() {
