+++ app/src/org/commcare/gis/EntityMapUtils.kt
@@ -1,6 +1,6 @@
 package org.commcare.gis
 
-import com.google.android.gms.maps.model.LatLng
+import com.mapbox.mapboxsdk.geometry.LatLng
 import org.commcare.CommCareApplication
 import org.commcare.activities.EntitySelectActivity
 import org.commcare.cases.entity.Entity
@@ -71,6 +71,20 @@ object EntityMapUtils {
             headers[i] = detail.fields[i].header.evaluate()
         }
         return headers
+    }
+
+    @JvmStatic
+    fun parseBoundaryCoords(boundaryCoords: String): ArrayList<LatLng> {
+        val latLngs = ArrayList<LatLng>()
+        if (boundaryCoords.isNotEmpty()) {
+            val list = boundaryCoords.split("\n")
+            list.filter { coord -> coord != "" }
+                    .map { coord ->
+                        val latLngArray = coord.split(",")
+                        LatLng(latLngArray[0].toDouble(), latLngArray[1].toDouble())
+                    }.toCollection(latLngs)
+        }
+        return latLngs
     }
 
 }
